requirements.txt
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.27.1
fastapi==0.109.1
orjson==3.9.7
pydantic==2.4.0
pydantic-settings==2.0.3
pydantic[email]
greenlet==2.0.2
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
asyncpg==0.28.0
psycopg2-binary==2.9.7
databases[postgresql]
redis==5.0.4
celery==5.4.0
flower==2.0.1
alembic==1.12.0
aiofiles==23.2.1
pathlib==1.0.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
aiosmtplib==2.0.2
jinja2==3.1.2
emails==0.6
docker==6.1.3
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
flake8==6.1.0